requirements.txt
Django>=3.2,<4.0
gunicorn>=20.1.0
reportlab>=3.6.0
PyMuPDF>=1.19.0
pdf2image>=1.16.0
Pillow>=8.0.0
nltk>=3.6.0
transformers>=4.0.0
torch>=1.7.0
python-dotenv>=0.19.0
django-environ>=0.4.5
django-crispy-forms>=1.12.0
opencv-python>=4.10.0.84
pytesseract>=0.3.13
pandas >= 2.2.3
